moved components to layout and elements folder
