+++ package.json
@@ -18,6 +18,7 @@
   "scripts": {
     "build-css": "tailwindcss build src/styles.css -o public/styles.css",
     "start": "node server.js",
+    "dev": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
