+++ src/routes/controls/index.js
@@ -1,15 +1,13 @@
 import React from 'react';
-import {Switch, Route} from 'react-router-dom';
+import { Route } from 'react-router-dom';
 import Group from '../../containers/controls/group';
 
 
 function ControlRoutes(){
     return(
-        <Switch>
              <Route path='/group'>
                 <Group />
             </Route>
-        </Switch>
     )
 }
